+++ lib/mod.ts
@@ -1,5 +1,6 @@
 export * from "./types.ts";
 export * from "./protocol.ts";
+export * from "./protocols.ts";
 export * from "./implementation.ts";
 export * from "./combine.ts";
-export * from "./reader.ts";
+export * from "./to-json.ts";
