+++ lib/protocols.ts
@@ -1,0 +1,4 @@
+export * as player  from "./protocols/player.ts";
+export * as scope  from "./protocols/scope.ts";
+
+
