--- scope/mod.ts
@@ -1,2 +1,0 @@
-export * from "./implementation.ts";
-export * from "./protocol.ts";
